refactor: remove client folder
