+++ backend/ws-server/src/main/java/org/linkwave/ws/websocket/dto/NewGroupChat.java
@@ -1,4 +1,4 @@
-package org.linkwave.ws.websocket.dto.client;
+package org.linkwave.ws.websocket.dto;
 
 import lombok.*;
